[PATRO23-7] all business logic move to KeycloakService from SignIn User
