+++ src/api/app/Program.cs
@@ -1,5 +1,5 @@
-using Intive.Patronage2023.Api;
 using Intive.Patronage2023.Api.Configuration;
+using Intive.Patronage2023.Api.Keycloak;
 using Intive.Patronage2023.Api.User;
 using Intive.Patronage2023.Modules.Example.Api;
 using Intive.Patronage2023.Shared.Abstractions;
